Rename thermal_network_matrix in all scripts that called functions in it
