+++ cea/plots/supply_system/main.py
@@ -107,7 +107,7 @@ def plots_main(locator, config):
 
 
 def preprocessing_run_thermal_network(config, locator, output_name_network, output_type_network):
-    from cea.technologies.thermal_network.thermal_network_matrix import thermal_network_main
+    from cea.technologies.thermal_network.thermal_network import thermal_network_main
     # configure thermal network (reduced simulation and create diagram of new network.
     config.restricted_to = None  # FIXME: remove this later
     network_name = output_name_network
